refactor(action): rename existing-name directive
